+++ lib/common/constants/app.dart
@@ -1,6 +1,6 @@
 // coverage:ignore-file
 
-const String version = '0.16.1';
+const String version = '0.17.0';
 
 /// https://m3.material.io/foundations/layout/applying-layout/window-size-classes
 const int maxScreenWidthCompact = 600;