+++ lib/filter/page/filter_list_page.dart
@@ -1,13 +1,17 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:go_router/go_router.dart';
-import 'package:ntodotxt/common/misc.dart' show PopScopeDrawer;
+import 'package:ntodotxt/common/misc.dart' show PopScopeDrawer, SnackBarHandler;
 import 'package:ntodotxt/common/widget/app_bar.dart';
 import 'package:ntodotxt/common/widget/chip.dart';
+import 'package:ntodotxt/common/widget/confirm_dialog.dart';
 import 'package:ntodotxt/common/widget/scroll_to_top.dart';
 import 'package:ntodotxt/filter/model/filter_model.dart';
 import 'package:ntodotxt/filter/state/filter_list_bloc.dart';
+import 'package:ntodotxt/filter/state/filter_list_event.dart';
 import 'package:ntodotxt/filter/state/filter_list_state.dart';
+import 'package:ntodotxt/setting/state/interaction_settings_cubit.dart';
+import 'package:ntodotxt/setting/state/interaction_settings_state.dart';
 import 'package:ntodotxt/todo/model/todo_model.dart' show Priority;
 
 class FilterListPage extends StatelessWidget {
@@ -131,15 +135,61 @@ class FilterListTile extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return ListTile(
-      key: ValueKey<int>(filter.id!),
-      title: Text(
-        filter.name,
-        maxLines: 1,
-        overflow: TextOverflow.ellipsis,
-      ),
-      subtitle: _buildSubtitle(),
-      onTap: () => context.pushNamed('filter-edit', extra: filter),
+    return BlocBuilder<InteractionSettingsCubit, InteractionSettingsState>(
+      buildWhen: (InteractionSettingsState previousState,
+              InteractionSettingsState state) =>
+          previousState.swipeRightActionEnabled !=
+          state.swipeRightActionEnabled,
+      builder: (BuildContext context, InteractionSettingsState state) {
+        return Dismissible(
+          key: ValueKey<int>(filter.id!),
+          direction: state.swipeRightActionEnabled
+              ? DismissDirection.startToEnd
+              : DismissDirection.none,
+          dismissThresholds: const {
+            DismissDirection.startToEnd: 0.5,
+            DismissDirection.endToStart: 0.5,
+          },
+          background: Container(
+            color: Theme.of(context).colorScheme.error, // red
+            alignment: Alignment.centerLeft,
+            padding: const EdgeInsets.only(left: 16.0),
+            child: Icon(Icons.delete),
+          ),
+          confirmDismiss: (DismissDirection direction) async {
+            // Delete
+            if (direction == DismissDirection.startToEnd) {
+              return await ConfirmationDialog.dialog(
+                context: context,
+                title: 'Delete filter',
+                message: 'Do you want to delete the filter?',
+                actionLabel: 'Delete',
+                cancelLabel: 'Cancel',
+              );
+            }
+            return true;
+          },
+          onDismissed: (DismissDirection direction) async {
+            if (direction == DismissDirection.startToEnd) {
+              // Delete
+              context.read<FilterListBloc>().add(
+                    FilterListFilterDeleted(filter: filter),
+                  );
+              SnackBarHandler.info(context, 'Filter has been deleted');
+            }
+          },
+          child: ListTile(
+            key: ValueKey<int>(filter.id!),
+            title: Text(
+              filter.name,
+              maxLines: 1,
+              overflow: TextOverflow.ellipsis,
+            ),
+            subtitle: _buildSubtitle(),
+            onTap: () => context.pushNamed('filter-edit', extra: filter),
+          ),
+        );
+      },
     );
   }