+++ lib/filter/state/filter_list_event.dart
@@ -1,4 +1,5 @@
 import 'package:equatable/equatable.dart';
+import 'package:ntodotxt/filter/model/filter_model.dart';
 
 sealed class FilterListEvent extends Equatable {
   const FilterListEvent();
@@ -13,4 +14,15 @@ final class FilterListSubscriped extends FilterListEvent {
 
 final class FilterListSynchronizationRequested extends FilterListEvent {
   const FilterListSynchronizationRequested();
+}
+
+final class FilterListFilterDeleted extends FilterListEvent {
+  final Filter filter;
+
+  const FilterListFilterDeleted({
+    required this.filter,
+  });
+
+  @override
+  List<Object?> get props => [filter];
 }