+++ test/setting/page/settings_page_test.dart
@@ -13,6 +13,7 @@ import 'package:ntodotxt/setting/controller/setting_controller.dart'
 import 'package:ntodotxt/setting/page/settings_page.dart' show SettingsPage;
 import 'package:ntodotxt/setting/repository/setting_repository.dart'
     show SettingRepository;
+import 'package:ntodotxt/setting/state/interaction_settings_cubit.dart';
 import 'package:ntodotxt/todo_file/state/todo_file_cubit.dart';
 import 'package:sqflite_common_ffi/sqflite_ffi.dart';
 
@@ -41,6 +42,11 @@ class SettingsPageBlocProvider extends StatelessWidget {
         RepositoryProvider<FilterRepository>(
           create: (BuildContext context) => FilterRepository(
             FilterController(dbController),
+          ),
+        ),
+        BlocProvider<InteractionSettingsCubit>(
+          create: (BuildContext context) => InteractionSettingsCubit(
+            repository: context.read<SettingRepository>(),
           ),
         ),
       ],