+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.17.0] - 2026-03-13
+
+### Added
+
+- Option to enable or disable swipe-right and swipe-left actions for list items #107
+- Option to delete filters via swipe-right action
+
 ## [0.16.1] - 2026-01-05
 
 ### Fixed
@@ -442,7 +449,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Intiial release
 
-[unreleased]: https://github.com/tmaegel/ntodotxt/compare/v0.16.1...HEAD
+[unreleased]: https://github.com/tmaegel/ntodotxt/compare/v0.17.0...HEAD
+[0.16.1]: https://github.com/tmaegel/ntodotxt/compare/v0.16.1...v0.17.0
 [0.16.1]: https://github.com/tmaegel/ntodotxt/compare/v0.16.0...v0.16.1
 [0.16.0]: https://github.com/tmaegel/ntodotxt/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/tmaegel/ntodotxt/compare/v0.14.2...v0.15.0